reorganizing cassie directory
