+++ examples/Cassie/systems/cassie_encoder.h
