+++ examples/Cassie/systems/sim_cassie_sensor_aggregator.h
@@ -4,18 +4,18 @@
 #include <string>
 #include <vector>
 
-#include "dairlib/lcmt_cassie_joint_out.hpp"
-#include "dairlib/lcmt_cassie_leg_out.hpp"
-#include "dairlib/lcmt_cassie_out.hpp"
-#include "dairlib/lcmt_cassie_pelvis_out.hpp"
-#include "dairlib/lcmt_elmo_out.hpp"
-#include "dairlib/lcmt_vectornav_out.hpp"
+#include "lcmtypes/_virtual_includes/lcmt_robot/dairlib/lcmt_cassie_joint_out.hpp"
+#include "lcmtypes/_virtual_includes/lcmt_robot/dairlib/lcmt_cassie_leg_out.hpp"
+#include "lcmtypes/_virtual_includes/lcmt_robot/dairlib/lcmt_cassie_out.hpp"
+#include "lcmtypes/_virtual_includes/lcmt_robot/dairlib/lcmt_cassie_pelvis_out.hpp"
+#include "lcmtypes/_virtual_includes/lcmt_robot/dairlib/lcmt_elmo_out.hpp"
+#include "lcmtypes/_virtual_includes/lcmt_robot/dairlib/lcmt_vectornav_out.hpp"
 #include "multibody/multibody_utils.h"
 #include "systems/framework/output_vector.h"
 #include "systems/framework/timestamped_vector.h"
 
-#include "drake/multibody/plant/multibody_plant.h"
-#include "drake/systems/framework/leaf_system.h"
+#include "external/drake/tools/install/libdrake/_virtual_includes/drake_shared_library/drake/multibody/plant/multibody_plant.h"
+#include "external/drake/tools/install/libdrake/_virtual_includes/drake_shared_library/drake/systems/framework/leaf_system.h"
 
 namespace dairlib {
 namespace systems {
