+++ examples/Cassie/test/input_supervisor_test.cc
@@ -1,4 +1,4 @@
-#include "examples/Cassie/input_supervisor.h"
+#include "examples/Cassie/systems/input_supervisor.h"
 
 #include <Eigen/Dense>
 #include <dairlib/lcmt_controller_failure.hpp>
